Always use ClassNameComparator when comparing classes. Adapt ClassNameComparator to correctly sort imports.
